Update prebid utils class to header bidding utils
