+++ hybid.demo/src/main/java/net/pubnative/lite/demo/ui/fragments/dfp/DFPMRectFragment.kt
@@ -43,7 +43,7 @@ import net.pubnative.lite.demo.util.ClipboardUtils
 import net.pubnative.lite.sdk.api.MRectRequestManager
 import net.pubnative.lite.sdk.api.RequestManager
 import net.pubnative.lite.sdk.models.Ad
-import net.pubnative.lite.sdk.utils.PrebidUtils
+import net.pubnative.lite.sdk.utils.HeaderBiddingUtils
 
 /**
  * Created by erosgarciaponte on 30.01.18.
@@ -105,12 +105,12 @@ class DFPMRectFragment : Fragment(), RequestManager.RequestListener {
     override fun onRequestSuccess(ad: Ad?) {
         val builder = PublisherAdRequest.Builder()
 
-        val keywordSet = PrebidUtils.getPrebidKeywordsSet(ad)
+        val keywordSet = HeaderBiddingUtils.getPrebidKeywordsSet(ad)
         for (key in keywordSet) {
             builder.addKeyword(key)
         }
 
-        val keywordBundle = PrebidUtils.getPrebidKeywordsBundle(ad)
+        val keywordBundle = HeaderBiddingUtils.getPrebidKeywordsBundle(ad)
         for (key in keywordBundle.keySet()) {
             builder.addCustomTargeting(key, keywordBundle.getString(key))
         }
